split out adjacent ashr shl pass
